Initial support for Mantle
This set of changes allows for generation of basic Mantle model files.  So far the +JSONKeyPathsByPropertyKey method has been fully implemented to allow for mapping between Obj-C properties and JSON keys.  There is no support for generating NSValueTransformer methods yet (this would probably call for additions to the schema), I haven't investigated this yet, but it should be doable.  Test coverage still needs to be updated to cover both JSONModel and Mantle usage.