Add array root object support
This change allows for the root of the JSON schema to be an array. This
means:

```
{
  "type": "array",
  "items": { ... }
}
```

is now totally valid and generates models properly.

This is an interesting case of model generation. In this case, we generate no
encompassing object, meaning we only generate the sub-object types in the array.
This makes sense in that, if your root is an array, you should be streaming
objects, not building a root object. Still, if you want to build a root object,
it's totally possible, but building in streaming is now about the same amount of
work, which is awesome.