Move from direct includes -> forward declarations
This commit adds support for circular dependencies in Nidyx. If you are
generating models in a case like:

```
A depends on B
B depends on C
C depends on A
```

Due to the fact that imports were used instead of forward declarations,
this would fail to compile. This commit solves this problem by moving to
forward declarations.